pyproject.toml
[project]
name = "research-project-template"
version = "0.1.0"
description = "KISS research project template built for Cursor IDE."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "<NAME>"},
]
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "hydra-submitit-launcher>=1.2.0",
    "ipykernel>=6.29.5",
    "loguru>=0.7.3",
    "wandb>=0.19.9",

]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]

[build]
target-dir = "build/dist"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 119
target-version = "py311"
